Depend on imglib2-mesh preview, not on net-imagej-mesh.
Still a WIP: several utility methods in TrackMate about meshes are now
in the imglib2-mesh dep.

+++ src/main/java/fiji/plugin/trackmate/visualization/bvv/BVVUtils.java
@@ -13,11 +13,11 @@ import ij.process.ImageProcessor;
 import ij.process.LUT;
 import net.imagej.ImgPlus;
 import net.imagej.axis.Axes;
-import net.imagej.mesh.Mesh;
-import net.imagej.mesh.Meshes;
-import net.imagej.mesh.nio.BufferMesh;
-import net.imagej.mesh.obj.transform.TranslateMesh;
 import net.imglib2.img.display.imagej.ImgPlusViews;
+import net.imglib2.mesh.Mesh;
+import net.imglib2.mesh.Meshes;
+import net.imglib2.mesh.impl.nio.BufferMesh;
+import net.imglib2.mesh.view.TranslateMesh;
 import net.imglib2.type.Type;
 import net.imglib2.type.numeric.ARGBType;
 
@@ -30,7 +30,7 @@ public class BVVUtils
 		{
 			final SpotMesh sm = ( SpotMesh ) spot;
 			final Mesh mesh = TranslateMesh.translate( sm.getMesh(), spot );
-			final BufferMesh bm = new BufferMesh( ( int ) mesh.vertices().size(), ( int ) mesh.triangles().size() );
+			final BufferMesh bm = new BufferMesh( mesh.vertices().size(), mesh.triangles().size() );
 			Meshes.copy( mesh, bm );
 			return new StupidMesh( bm );
 		}
